test4

+++ lib/features/logbook/log_controller.dart
@@ -3,12 +3,15 @@ import 'package:mongo_dart/mongo_dart.dart';
 import 'package:logbook_app_001/features/logbook/models/log_model.dart';
 import 'package:logbook_app_001/services/mongo_service.dart';
 import 'package:logbook_app_001/helpers/log_helper.dart';
+import 'package:logbook_app_001/services/access_control_service.dart'; 
 
 class LogController {
   final String username;
+  
+  final currentUser = (id: 'user_001', role: 'Anggota', teamId: 'team_polban_01');
+
   final ValueNotifier<List<LogModel>> logsNotifier = ValueNotifier([]);
   final ValueNotifier<List<LogModel>> filteredLogs = ValueNotifier([]);
-  
   final ValueNotifier<bool> isLoading = ValueNotifier(false);
 
   LogController(this.username);
@@ -40,7 +43,10 @@ class LogController {
       description: desc,
       date: DateTime.now(),
       category: category,
+      authorId: currentUser.id,   
+      teamId: currentUser.teamId,  
     );
+    
     try {
       await MongoService().insertLog(newLog);
       logsNotifier.value = [...logsNotifier.value, newLog];
@@ -52,17 +58,22 @@ class LogController {
 
   Future<void> updateLog(int index, String title, String desc, String category) async {
     final logToUpdate = filteredLogs.value[index];
+    
     final updatedLog = LogModel(
       id: logToUpdate.id, 
       title: title,
       description: desc,
-      date: DateTime.now(),
+      date: DateTime.now(), 
       category: category,
+      authorId: logToUpdate.authorId, 
+      teamId: logToUpdate.teamId,     
     );
+
     try {
       await MongoService().updateLog(updatedLog);
       final currentLogs = List<LogModel>.from(logsNotifier.value);
       final realIndex = logsNotifier.value.indexOf(logToUpdate);
+      
       if (realIndex != -1) {
         currentLogs[realIndex] = updatedLog;
         logsNotifier.value = currentLogs;
@@ -75,13 +86,22 @@ class LogController {
 
   Future<void> removeLog(int index) async {
     final logToRemove = filteredLogs.value[index];
+    
     try {
+      bool canDelete = AccessControlService.canPerform(currentUser.role, 'delete', isOwner: logToRemove.authorId == currentUser.id);
+
+      if (!canDelete) {
+        await LogHelper.writeLog("SECURITY: Unauthorized delete attempt by ${currentUser.id}", level: 1);
+        return;
+      } 
+
       if (logToRemove.id != null) {
         await MongoService().deleteLog(logToRemove.id!);
         final currentLogs = List<LogModel>.from(logsNotifier.value);
-        currentLogs.remove(logToRemove);
+        currentLogs.removeWhere((item) => item.id == logToRemove.id);
+        
         logsNotifier.value = currentLogs;
-        filteredLogs.value = logsNotifier.value;
+        filteredLogs.value = List.from(currentLogs);
       }
     } catch (e) {
       LogHelper.writeLog("Gagal hapus log: $e", level: 1);
